refactor(collavre): extract CreativeTreeSerializer from CreativesController
Move the simple-mode flat serialization (serialize_creatives plus its
children_presence_set and mask_unreachable_crumbs helpers) into
Collavre::Creatives::CreativeTreeSerializer. The controller keeps a thin
serialize_creatives delegator so the picker test seam is preserved and the
JSON response shape is byte-for-byte identical.
